+++ x/staking/abci_test.go
@@ -25,8 +25,6 @@ func TestItUpdatesTheValidatorSetBasedOnPendingValidators(t *testing.T) {
 	require.NoError(t, err)
 
 	validator1, err := stakingtypes.NewValidator(
-		"validator 1",
-		types.NewDec(1234),
 		valAddr1,
 		pubKey,
 	)
@@ -109,8 +107,6 @@ func TestItRemovesFromTheValidatorSetWhenInRemovingQueue(t *testing.T) {
 			require.NoError(t, err)
 
 			validator1, err := stakingtypes.NewValidator(
-				"validator 1",
-				types.NewDec(1234),
 				valAddr1,
 				pubKey,
 			)
@@ -194,8 +190,6 @@ func TestItIncludesItBackToValidatorSetOnceReactivatingIt(t *testing.T) {
 			require.NoError(t, err)
 
 			validator1, err := stakingtypes.NewValidator(
-				"validator 1",
-				types.NewDec(1234),
 				valAddr1,
 				pubKey,
 			)