+++ x/staking/client/cli/query.go
@@ -13,11 +13,6 @@ import (
 	"github.com/spf13/pflag"
 )
 
-const (
-	FlagValAddr = "val-addr"
-	FlagAddr    = "addr"
-)
-
 // GetQueryCmd returns the cli query commands for this module
 func GetQueryCmd() *cobra.Command {
 	queryCmd := &cobra.Command{
@@ -30,15 +25,71 @@ func GetQueryCmd() *cobra.Command {
 
 	queryCmd.AddCommand(
 		GetCmdQueryValidator(),
+		GetCmdQueryValidators(),
 	)
 
 	return queryCmd
 }
 
 // GetCmdQueryValidator the query delegation command.
+func GetCmdQueryValidators() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "validators",
+		Short: "Query a validators by specific filter  [--addr || --val-addr || --moniker || --status || --pubkey || --proposer ]",
+		RunE: func(cmd *cobra.Command, args []string) error {
+			clientCtx := client.GetClientContextFromCmd(cmd)
+
+			pubKey, _ := cmd.Flags().GetString(FlagPubKey)
+			proposer, _ := cmd.Flags().GetString(FlagProposer)
+			status, _ := cmd.Flags().GetString(FlagStatus)
+			valAddr, _ := cmd.Flags().GetString(FlagValAddr)
+			addr, _ := cmd.Flags().GetString(FlagAddr)
+			moniker, _ := cmd.Flags().GetString(FlagMoniker)
+
+			pageReq, err := client.ReadPageRequest(cmd.Flags())
+			if err != nil {
+				return err
+			}
+
+			request := &stakingtypes.ValidatorsRequest{
+				Address:    addr,
+				Valkey:     valAddr,
+				Pubkey:     pubKey,
+				Moniker:    moniker,
+				Status:     status,
+				Proposer:   proposer,
+				Pagination: pageReq,
+			}
+
+			queryClient := stakingtypes.NewQueryClient(clientCtx)
+			res, err := queryClient.Validators(context.Background(), request)
+			if err != nil {
+				return err
+			}
+
+			return clientCtx.PrintProto(res)
+
+		},
+	}
+
+	flags.AddQueryFlagsToCmd(cmd)
+
+	cmd.Flags().String(FlagAddr, "", "the addres in AccAddress format.")
+	cmd.Flags().String(FlagValAddr, "", "the addres in ValAddress format.")
+	cmd.Flags().String(FlagMoniker, "", "the moniker")
+	cmd.Flags().String(FlagPubKey, "", "the pubKey")
+	cmd.Flags().String(FlagProposer, "", "the proposer")
+	cmd.Flags().String(FlagStatus, "", "the status")
+
+	flags.AddPaginationFlagsToCmd(cmd, "customstaking")
+
+	return cmd
+}
+
+// GetCmdQueryValidator the query delegation command.
 func GetCmdQueryValidator() *cobra.Command {
 	cmd := &cobra.Command{
-		Use:   "validator [--addr || --val-addr || --flagMoniker] ",
+		Use:   "validator [--addr || --val-addr || --moniker] ",
 		Short: "Query a validator based on address",
 		RunE: func(cmd *cobra.Command, args []string) error {
 			clientCtx := client.GetClientContextFromCmd(cmd)