+++ x/staking/keeper/msg_server.go
@@ -48,23 +48,34 @@ func (k msgServer) ClaimValidator(goCtx context.Context, msg *types.MsgClaimVali
 		return nil, types.ErrValidatorAlreadyClaimed
 	}
 
-	_, err = k.keeper.GetValidatorByMoniker(ctx, msg.Moniker)
+	moniker := strings.Trim(msg.Moniker, " ")
+	_, err = k.keeper.GetValidatorByMoniker(ctx, moniker)
 	if err == nil {
 		return nil, types.ErrValidatorMonikerExists
 	}
 
-	validator, err := types.NewValidator(strings.Trim(msg.Moniker, " "), msg.Commission, msg.ValKey, pk)
+	validator, err := types.NewValidator(msg.ValKey, pk)
 	if err != nil {
 		return nil, err
 	}
 
 	k.keeper.AddPendingValidator(ctx, validator)
 
+	// register identity record moniker field when claiming a validator
+	err = k.govKeeper.RegisterIdentityRecords(ctx, sdk.AccAddress(validator.ValKey), []govtypes.IdentityInfoEntry{
+		{
+			Key:  "moniker",
+			Info: moniker,
+		},
+	})
+	if err != nil {
+		return nil, err
+	}
+
 	ctx.EventManager().EmitEvent(
 		sdk.NewEvent(
 			types.EventTypeClaimValidator,
 			sdk.NewAttribute(types.AttributeKeyMoniker, msg.Moniker),
-			sdk.NewAttribute(types.AttributeKeyCommission, msg.Commission.String()),
 			sdk.NewAttribute(types.AttributeKeyValKey, msg.ValKey.String()),
 			sdk.NewAttribute(types.AttributeKeyPubKey, msg.PubKey.String()),
 		),